pyflink/requirements.txt
Faker
psycopg2-binary
tzdata